test: merge battle control state tests
