+++ tests/helpers/classicBattle/controlState.test.js
@@ -1,30 +1,42 @@
-import { describe, it, expect, vi, beforeEach } from "vitest";
+import { describe, it, expect, vi, beforeEach, afterEach } from "vitest";
 import "./commonMocks.js";
+import { createRoundMessage, createTimerNodes } from "./domUtils.js";
 import { createStateManager } from "../../../src/helpers/classicBattle/stateManager.js";
 import {
   waitingForPlayerActionEnter,
   waitingForPlayerActionExit
 } from "../../../src/helpers/classicBattle/orchestratorHandlers.js";
 import { ClassicBattleView } from "../../../src/helpers/classicBattle/view.js";
+
 const currentFlags = { statHotkeys: { enabled: false } };
+
+function mockQuitMatch() {
+  const msg = document.getElementById("round-message");
+  if (msg) msg.textContent = "quit";
+  window.location.href = "http://localhost/index.html";
+}
+
 vi.mock("../../../src/helpers/featureFlags.js", () => ({
   isEnabled: (flag) => currentFlags[flag]?.enabled ?? false,
   featureFlagsEmitter: new EventTarget(),
   initFeatureFlags: vi.fn()
 }));
 vi.mock("../../../src/helpers/setupScoreboard.js", () => ({ setupScoreboard: vi.fn() }));
-vi.mock("../../../src/helpers/classicBattle/quitButton.js", () => ({ initQuitButton: vi.fn() }));
 vi.mock("../../../src/helpers/classicBattle/skipHandler.js", () => ({ skipCurrentPhase: vi.fn() }));
 vi.mock("../../../src/helpers/classicBattle/interruptHandlers.js", () => ({
   initInterruptHandlers: vi.fn()
 }));
+vi.mock("../../../src/helpers/classicBattle/timerService.js", async () => {
+  const actual = await vi.importActual("../../../src/helpers/classicBattle/timerService.js");
+  return { ...actual, onNextButtonClick: vi.fn() };
+});
 vi.mock("../../../src/helpers/classicBattle/uiHelpers.js", async () => {
   const actual = await vi.importActual("../../../src/helpers/classicBattle/uiHelpers.js");
   return {
     ...actual,
     watchBattleOrientation: vi.fn(),
     registerRoundStartErrorHandler: vi.fn(),
-    setupNextButton: vi.fn(),
+    setupNextButton: actual.setupNextButton,
     applyStatLabels: vi.fn().mockResolvedValue(),
     setBattleStateBadgeEnabled: vi.fn(),
     applyBattleFeatureFlags: vi.fn(),
@@ -39,11 +51,86 @@ vi.mock("../../../src/helpers/tooltip.js", () => ({ initTooltips: vi.fn().mockRe
 vi.mock("../../../src/helpers/setupBottomNavbar.js", () => ({}));
 vi.mock("../../../src/helpers/setupDisplaySettings.js", () => ({}));
 vi.mock("../../../src/helpers/setupSvgFallback.js", () => ({}));
-vi.mock("../../../src/helpers/setupClassicBattleHomeLink.js", () => ({}));
 
-describe("classicBattle stat button state", () => {
+vi.mock("../../../src/helpers/classicBattle/quitModal.js", () => ({
+  quitMatch: mockQuitMatch
+}));
+
+describe("classicBattle battle control state", () => {
   beforeEach(() => {
-    document.body.innerHTML = '<div id="stat-buttons"><button data-stat="power"></button></div>';
+    document.body.innerHTML = "";
+    createRoundMessage();
+    const { nextButton } = createTimerNodes();
+    nextButton.disabled = true;
+    const quitBtn = document.createElement("button");
+    quitBtn.id = "quit-match-button";
+    const header = document.createElement("header");
+    const statButtons = document.createElement("div");
+    statButtons.id = "stat-buttons";
+    statButtons.innerHTML = '<button data-stat="power"></button>';
+    document.body.append(quitBtn, header, statButtons);
+    window.location.href = "http://localhost/src/pages/battleJudoka.html";
+    vi.resetModules();
+  });
+
+  afterEach(() => {
+    vi.restoreAllMocks();
+  });
+
+  it("enable/disable helpers toggle button state", async () => {
+    const { disableNextRoundButton, enableNextRoundButton } = await import(
+      "../../../src/helpers/classicBattle.js"
+    );
+    const btn = document.getElementById("next-button");
+    disableNextRoundButton();
+    expect(btn.disabled).toBe(true);
+    expect(btn.dataset.nextReady).toBeUndefined();
+    enableNextRoundButton();
+    expect(btn.disabled).toBe(false);
+    expect(btn.dataset.nextReady).toBe("true");
+  });
+
+  it("resetBattleUI replaces Next button and reattaches click handler", async () => {
+    const timerSvc = await import("../../../src/helpers/classicBattle/timerService.js");
+    const { resetBattleUI } = await import("../../../src/helpers/classicBattle/uiHelpers.js");
+    const btn = document.getElementById("next-button");
+    btn.dataset.nextReady = "true";
+    resetBattleUI();
+    const cloned = document.getElementById("next-button");
+    expect(cloned).not.toBe(btn);
+    expect(cloned.disabled).toBe(true);
+    expect(cloned.dataset.nextReady).toBeUndefined();
+    cloned.dispatchEvent(new MouseEvent("click"));
+    expect(timerSvc.onNextButtonClick).toHaveBeenCalledTimes(1);
+  });
+
+  it("quit button triggers quitMatch", async () => {
+    const { createBattleStore } = await import(
+      "../../../src/helpers/classicBattle/roundManager.js"
+    );
+    const { initQuitButton } = await import("../../../src/helpers/classicBattle/quitButton.js");
+    window.battleStore = createBattleStore();
+    initQuitButton(window.battleStore, { quitMatch: mockQuitMatch });
+    document.getElementById("quit-match-button").click();
+    expect(document.getElementById("round-message").textContent).toBe("quit");
+    expect(window.location.href).toBe("http://localhost/index.html");
+  });
+
+  it("home link invokes quitMatch", async () => {
+    const header = document.querySelector("header");
+    const homeLink = document.createElement("a");
+    homeLink.href = "../../index.html";
+    homeLink.dataset.testid = "home-link";
+    header.appendChild(homeLink);
+
+    await import("../../../src/helpers/setupClassicBattleHomeLink.js");
+    const { createBattleStore } = await import(
+      "../../../src/helpers/classicBattle/roundManager.js"
+    );
+    window.battleStore = createBattleStore();
+    homeLink.click();
+    expect(document.getElementById("round-message").textContent).toBe("quit");
+    expect(window.location.href).toBe("http://localhost/index.html");
   });
 
   it("enables stat buttons only while waiting for player action", async () => {
